pyproject.toml
[project]
name = "plasma-mcp-server"
version = "0.1.0"
description = "MCP Server and AI Agent for Plasma UI Kit"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "whoosh>=2.7.4",
    "markdown-it-py>=3.0.0",
    "gigachat>=0.1.18",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
